separate book parsing from tune book
